requirements-gadi.txt
absl-py==0.15.0
astunparse==1.6.3
attrs==20.3.0
backcall==0.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.14.6
charset-normalizer==3.3.2
click==8.1.7
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
distro==1.5.0
docker-pycreds==0.4.0
einops==0.4.1
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
hypothesis==6.8.3
idna==3.7
importlib_metadata==7.1.0
iniconfig==1.1.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.4.0
ninja==1.10.0.post2
numpy==1.22.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
platformdirs==4.2.1
pluggy==0.13.1
prompt-toolkit==3.0.18
protobuf==3.15.0
psutil==5.9.8
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-build==0.11.1
scipy==1.6.2
sentry-sdk==2.1.1
setproctitle==1.3.3
shortio==0.1.0
sio==0.1.0
six==1.15.0
smmap==5.0.1
sortedcontainers==2.3.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
toml==0.10.2
torch==1.9.0a0+gitd69c22d
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==2.2.1
wandb==0.17.0
wcwidth==0.2.5
Werkzeug==3.0.3
wrapt==1.12.1
zipp==3.18.1

